Allow all ip addresses that start with "172."
Hey Voyz! 

Thank you so much for this awesome repo, it's really a game-changer for me :)

I have a small issue with one of my web services that run on a private network on my docker-compose.
The IP address of web services always changing, but the one source of truth is that it always starts with 172.

Is there a reason why the specific address of `172.17.0.*` was specified in the allow list?
Is there a security risk in accepting all addresses that start with 172?